pyproject.toml
[tool.poetry]
name = "qol_airflow"
version = "0.1.0"
description = "Airflow for QoL project"
authors = ["Chayapol Parisonyu <chayapol.par@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
apache-airflow = "^2.3.0"
apache-airflow-providers-amazon = "^3.4.0"
pyproj = "^3.3.1"
networkx = "^2.8"
Shapely = "^1.8.2"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
